python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unicefdata"
version = "2.1.0"
description = "Python library for downloading UNICEF indicators via SDMX API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    { name = "Joao Pedro Azevedo", email = "jpazevedo@unicef.org" }
]
keywords = ["unicef", "sdmx", "indicators", "data", "api", "statistics", "sdg", "development"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
excel = ["openpyxl>=3.0.0", "xlrd>=2.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://jpazvd.github.io/"
Documentation = "https://github.com/unicef-drp/unicefData/tree/main/python"
Repository = "https://github.com/unicef-drp/unicefData"
Changelog = "https://github.com/unicef-drp/unicefData/blob/main/python/CHANGELOG.md"
"Bug Tracker" = "https://github.com/unicef-drp/unicefData/issues"

[tool.setuptools.packages.find]
include = ["unicefdata*"]

[tool.setuptools.package-data]
unicefdata = ["metadata/**/*.yaml"]
